+++ src/shillelagh/adapters/api/gsheets/lib.py
@@ -14,11 +14,13 @@ from shillelagh.adapters.api.gsheets.fields import (
     GSheetsBoolean,
     GSheetsDate,
     GSheetsDateTime,
+    GSheetsDuration,
     GSheetsField,
     GSheetsNumber,
     GSheetsString,
     GSheetsTime,
 )
+from shillelagh.adapters.api.gsheets.parsing.date import infer_column_type
 from shillelagh.adapters.api.gsheets.types import SyncMode
 from shillelagh.adapters.api.gsheets.typing import (
     QueryResultsCell,
@@ -46,10 +48,10 @@ def get_field(
     """
     Return a Shillelagh ``Field`` from a Google Chart API results column.
     """
-    # Fix for GSheets return an incorrect type. We should be able to detect the type based
-    # on the pattern, instead of relying on the return type.
-    if col["type"] == "datetime" and col.get("pattern") == "h:mm:ss am/pm":
-        col["type"] = "timeofday"
+    # GSheets returns type ``datetime`` for timestamps, but also for time of day and
+    # durations. We need to tokenize the pattern in order to figure out the correct type.
+    if col["type"] == "datetime" and "pattern" in col:
+        col["type"] = infer_column_type(col["pattern"])
 
     type_map: Dict[str, Tuple[Type[GSheetsField], List[Type[Filter]]]] = {
         "string": (GSheetsString, [Range, Equal, NotEqual, Like, IsNull, IsNotNull]),
@@ -58,6 +60,7 @@ def get_field(
         "date": (GSheetsDate, [Range, Equal, NotEqual, IsNull, IsNotNull]),
         "datetime": (GSheetsDateTime, [Range, Equal, NotEqual, IsNull, IsNotNull]),
         "timeofday": (GSheetsTime, [Range, Equal, NotEqual, IsNull, IsNotNull]),
+        "duration": (GSheetsDuration, [Range, Equal, NotEqual, IsNull, IsNotNull]),
     }
     class_, filters = type_map.get(
         col["type"],