+++ CHANGELOG.rst
@@ -8,7 +8,7 @@ Next
 Version 1.2.6 - 2023-07-20
 ==========================
 
-- Add support for querying durations in Google Sheets ()
+- Add support for querying durations in Google Sheets (#374)
 
 Version 1.2.5 - 2023-07-14
 ==========================